型情報をリファクタ
